🛠️ Refactor code

+++ src/Application/APIKeys/Commands/DeleteAPIKey/DeleteAPIKey.cs
@@ -1,4 +1,5 @@
-﻿using AIExtensionsCenter.Application.Common.Interfaces;
+﻿using AIExtensionsCenter.Application.Common.Exceptions;
+using AIExtensionsCenter.Application.Common.Interfaces;
 using AIExtensionsCenter.Domain.Entities;
 using Ardalis.GuardClauses;
 
@@ -19,16 +20,20 @@ public class DeleteAPIKeyCommandValidator : AbstractValidator<DeleteAPIKeyComman
 public class DeleteAPIKeyCommandHandler : IRequestHandler<DeleteAPIKeyCommand>
 {
     private readonly IApplicationDbContext _context;
+    private readonly IUser _user;
 
-    public DeleteAPIKeyCommandHandler(IApplicationDbContext context)
+    public DeleteAPIKeyCommandHandler(IApplicationDbContext context, IUser user)
     {
         _context = context;
+        _user = user;
     }
 
     public async Task Handle(DeleteAPIKeyCommand request, CancellationToken cancellationToken)
     {
         APIKey? apikey = await _context.APIKeys.FirstOrDefaultAsync(x => x.Id == request.Id, cancellationToken);
         Guard.Against.NotFound(request.Id, apikey);
+
+        if (apikey.UserId != _user.Id) throw new ForbiddenAccessException();
 
         _context.APIKeys.Remove(apikey);
         await _context.SaveChangesAsync(cancellationToken);
