+++ src/Application/Licenses/Commands/DeleteLicense/DeleteLicense.cs
@@ -1,4 +1,5 @@
-﻿using AIExtensionsCenter.Application.Common.Interfaces;
+﻿using AIExtensionsCenter.Application.Common.Exceptions;
+using AIExtensionsCenter.Application.Common.Interfaces;
 using AIExtensionsCenter.Domain.Entities;
 using Ardalis.GuardClauses;
 
@@ -19,16 +20,19 @@ public class DeleteLicenseCommandValidator : AbstractValidator<DeleteLicenseComm
 public class DeleteLicenseCommandHandler : IRequestHandler<DeleteLicenseCommand>
 {
     private readonly IApplicationDbContext _context;
-
-    public DeleteLicenseCommandHandler(IApplicationDbContext context)
+    private readonly IUser _user;
+    public DeleteLicenseCommandHandler(IApplicationDbContext context, IUser user)
     {
         _context = context;
+        _user = user;
     }
 
     public async Task Handle(DeleteLicenseCommand request, CancellationToken cancellationToken)
     {
         License? license = await _context.Licenses.FirstOrDefaultAsync(x => x.Id == request.Id, cancellationToken);
         Guard.Against.NotFound(request.Id, license);
+
+        if (license.UserId != _user.Id) throw new ForbiddenAccessException();
 
         _context.Licenses.Remove(license);
         await _context.SaveChangesAsync(cancellationToken);
